use different xpath to specify wether guide curves shall be converted
previously a user could add the xpath "cpacs/header/cpacsVersionGuideCurveInterp" to specify the interpretation of the guide curve profile points. If this textElement was "3", the guide curves are kept the way they are.

Now the user can add a Boolean xpath "cpacs/header/update/cpacs2to3/configuration/convertGuideCurves" which is more intuitive and extendable to other conversion customizations

addresses #15